+++ src/updater/imp.rs
@@ -1,14 +1,18 @@
-use super::*;
+use super::{
+    anyhow, env, env_logger, remove_file, DateTime, PathBuf, Receiver, RefCell, Releaser, Result,
+    Url, Utc, Version, UPDATE_INTERVAL,
+};
+use crate::Updater;
 use std::cell::Cell;
 use std::cell::Ref;
 use std::cell::RefMut;
+use std::path::Path;
 use std::sync::mpsc;
-use crate::Updater;
 
 pub(super) const LATEST_UPDATE_INFO_CACHE_FN_ASYNC: &str = "last_check_status_async.json";
 
 // Payload that the worker thread will send back
-type ReleasePayloadResult = Result<Option<UpdateInfo>, Error>;
+type ReleasePayloadResult = Result<Option<UpdateInfo>>;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub(super) struct UpdaterState {
@@ -65,10 +69,7 @@ pub(super) struct UpdateInfo {
     pub fetched_at: Option<DateTime<Utc>>,
 
     // Link to use to download the above version
-    #[serde(with = "url_serde")]
     pub downloadable_url: Url,
-
-    _priv: (),
 }
 
 impl UpdateInfo {
@@ -77,7 +78,6 @@ impl UpdateInfo {
             version: v,
             fetched_at: None,
             downloadable_url: url,
-            _priv: (),
         }
     }
 
@@ -115,7 +115,7 @@ impl<T> Updater<T>
 where
     T: Releaser + Send + 'static,
 {
-    pub(super) fn load_or_new(r: T) -> Result<Self, Error> {
+    pub(super) fn load_or_new(r: T) -> Result<Self> {
         let _ = env_logger::try_init();
         if let Ok(mut saved_state) = Self::load() {
             // Use the version that workflow reports through environment variable
@@ -130,7 +130,7 @@ where
             })
         } else {
             let current_version = env::workflow_version()
-                .map_or_else(|| Ok(Version::from((0, 0, 0))), |v| Version::parse(&v))?;
+                .map_or_else(|| Ok(Version::new(0, 0, 0)), |v| Version::parse(&v))?;
             let state = UpdaterState {
                 current_version,
                 last_check: Cell::new(None),
@@ -163,18 +163,18 @@ where
         self.state.update_interval = t;
     }
 
-    fn load() -> Result<UpdaterState, Error> {
+    fn load() -> Result<UpdaterState> {
         let data_file_path = Self::build_data_fn()?;
         crate::Data::load_from_file(data_file_path)
-            .ok_or_else(|| err_msg("cannot load cached state of updater"))
+            .ok_or_else(|| anyhow!("cannot load cached state of updater"))
     }
 
     // Save updater's state
-    pub(super) fn save(&self) -> Result<(), Error> {
+    pub(super) fn save(&self) -> Result<()> {
         let data_file_path = Self::build_data_fn()?;
-        crate::Data::save_to_file(&data_file_path, &self.state).or_else(|e| {
-            let _ = remove_file(data_file_path);
-            Err(e)
+        crate::Data::save_to_file(&data_file_path, &self.state).map_err(|e| {
+            let _r = remove_file(data_file_path);
+            e
         })
     }
 
@@ -182,14 +182,14 @@ where
         &self,
         tx: mpsc::Sender<ReleasePayloadResult>,
         p: PathBuf,
-    ) -> Result<(), Error> {
+    ) -> Result<()> {
         use std::thread;
 
         let releaser = (*self.releaser.borrow()).clone();
 
         thread::Builder::new().spawn(move || {
             debug!("other thread: starting in updater thread");
-            let talk_to_mother = || -> Result<(), Error> {
+            let talk_to_mother = || -> Result<()> {
                 let (v, url) = releaser.latest_release()?;
                 let mut info = UpdateInfo::new(v, url);
                 info.set_fetched_at(Utc::now());
@@ -212,21 +212,21 @@ where
 
     // write version of latest avail. release (if any) to a cache file
     pub(super) fn write_last_check_status(
-        p: &PathBuf,
+        p: &Path,
         updater_info: &Option<UpdateInfo>,
-    ) -> Result<(), Error> {
-        crate::Data::save_to_file(p, updater_info).or_else(|e| {
-            let _ = remove_file(p);
-            Err(e)
+    ) -> Result<()> {
+        crate::Data::save_to_file(p, updater_info).map_err(|e| {
+            let _r = remove_file(p);
+            e
         })
     }
 
     // read version of latest avail. release (if any) from a cache file
-    pub(super) fn read_last_check_status(p: &PathBuf) -> Result<Option<UpdateInfo>, Error> {
-        crate::Data::load_from_file(p).ok_or_else(|| err_msg("no data in given path"))
+    pub(super) fn read_last_check_status(p: &Path) -> Result<Option<UpdateInfo>> {
+        crate::Data::load_from_file(p).ok_or_else(|| anyhow!("no data in given path"))
     }
 
-    pub(super) fn build_data_fn() -> Result<PathBuf, Error> {
+    pub(super) fn build_data_fn() -> Result<PathBuf> {
         let workflow_name = env::workflow_name()
             .unwrap_or_else(|| "YouForgotTo/フ:NameYourOwnWork}flowッ".to_string())
             .chars()
@@ -235,40 +235,40 @@ where
 
         env::workflow_cache()
             .ok_or_else(|| {
-                err_msg("missing env variable for cache dir. forgot to set workflow bundle id?")
+                anyhow!("missing env variable for cache dir. forgot to set workflow bundle id?")
             })
             .and_then(|mut data_path| {
                 env::workflow_uid()
-                    .ok_or_else(|| err_msg("missing env variable for uid"))
-                    .and_then(|ref uid| {
+                    .ok_or_else(|| anyhow!("missing env variable for uid"))
+                    .map(|ref uid| {
                         let filename = [uid, "-", workflow_name.as_str(), "-updater.json"].concat();
                         data_path.push(filename);
 
-                        Ok(data_path)
+                        data_path
                     })
             })
     }
 
-    pub(super) fn update_ready_async(&self, try_flag: bool) -> Result<bool, Error> {
+    pub(super) fn update_ready_async(&self, try_flag: bool) -> Result<bool> {
         self.state
             .worker_state
             .borrow()
             .as_ref()
-            .ok_or_else(|| err_msg("you need to use init() method first."))
+            .ok_or_else(|| anyhow!("you need to use init() method first."))
             .and_then(|mpsc| {
                 if mpsc.recvd_payload.borrow().is_none() {
                     // No payload received yet, try to talk to worker thread
                     mpsc.rx
                         .borrow()
                         .as_ref()
-                        .ok_or_else(|| err_msg("you need to use init() correctly!"))
+                        .ok_or_else(|| anyhow!("you need to use init() correctly!"))
                         .and_then(|rx| {
                             let rr = if try_flag {
                                 // don't block while trying to receive
-                                rx.try_recv().map_err(|e| err_msg(e.to_string()))
+                                rx.try_recv().map_err(|e| anyhow!(e.to_string()))
                             } else {
                                 // block while waiting to receive
-                                rx.recv().map_err(|e| err_msg(e.to_string()))
+                                rx.recv().map_err(|e| anyhow!(e.to_string()))
                             };
                             rr.and_then(|msg| {
                                 let msg_status = msg.map(|update_info| {
@@ -281,7 +281,7 @@ where
                                                 || self.last_check().as_ref().unwrap()
                                                     < fetched_time
                                             {
-                                                self.set_last_check(*fetched_time)
+                                                self.set_last_check(*fetched_time);
                                             }
                                         })
                                     });
@@ -301,16 +301,14 @@ where
             .avail_release
             .borrow()
             .as_ref()
-            .map(|release| *self.current_version() < release.version)
-            .unwrap_or(false))
+            .map_or(false, |release| *self.current_version() < release.version))
     }
 
     #[allow(dead_code)]
-    pub(super) fn _update_ready_async(&self) -> Result<bool, Error> {
+    #[deprecated(note = "update_ready_async is deprecated. use init()")]
+    pub(super) fn _update_ready_async(&self) -> Result<bool> {
         let worker_state = self.state.worker_state.borrow();
-        if worker_state.is_none() {
-            panic!("you need to use init first")
-        };
+        assert!(worker_state.is_some(), "you need to use init first");
 
         let mpsc = worker_state.as_ref().expect("no worker_state");
         if mpsc.recvd_payload.borrow().is_none() {
@@ -324,12 +322,12 @@ where
                     *self.state.avail_release.borrow_mut() = update_info.clone();
                     *mpsc.recvd_payload.borrow_mut() = Some(Ok(update_info.clone()));
                 } else {
-                    return Err(err_msg(format!("{:?}", msg.as_ref().unwrap_err())));
+                    return Err(anyhow!(format!("{:?}", msg.as_ref().unwrap_err())));
                 }
                 self.save()?;
             } else {
                 eprintln!("{:?}", rr);
-                return Err(err_msg(format!("{:?}", rr)));
+                return Err(anyhow!(format!("{:?}", rr)));
             }
         }
         if let Some(ref updater_info) = *self.state.avail_release.borrow() {
@@ -344,7 +342,8 @@ where
     }
 
     #[allow(dead_code)]
-    pub(super) fn _update_ready_sync(&self) -> Result<bool, Error> {
+    #[deprecated(note = "update_ready_sync is deprecated. use init()")]
+    pub(super) fn _update_ready_sync(&self) -> Result<bool> {
         // A None value for last_check indicates that workflow is being run for first time.
         // Thus we update last_check to now and just save the updater state without asking
         // Releaser to do a remote call/check for us since we assume that user just downloaded
@@ -356,7 +355,7 @@ where
 
         // make a network call to see if a newer version is avail.
         // save the result of call to cache file.
-        let ask_releaser_for_update = || -> Result<bool, Error> {
+        let ask_releaser_for_update = || -> Result<bool> {
             let (v, url) = self.releaser.borrow().latest_release()?;
             let update_avail = *self.current_version() < v;
 
@@ -386,9 +385,10 @@ where
         } else {
             Self::read_last_check_status(&p)
                 .map(|last_check_status| {
-                    last_check_status
-                        .map(|last_update_info| *self.current_version() < last_update_info.version)
-                        .unwrap_or(false)
+                    last_check_status.map_or(false, |last_update_info| {
+                        *self.current_version() < last_update_info.version
+                    })
+                    // .unwrap_or(false)
                 })
                 .or(Ok(false))
         }