+++ src/updater/tests.rs
@@ -1,5 +1,5 @@
 use self::releaser::tests::setup_mock_server;
-#[cfg(not(feature = "ci"))]
+// #[cfg(not(feature = "ci"))]
 use self::releaser::GithubReleaser;
 use self::releaser::MOCK_RELEASER_REPO_NAME;
 use super::*;
@@ -32,10 +32,7 @@ fn it_ignores_saved_version_after_an_upgrade_async() {
         updater.set_interval(0);
         updater.init().expect("couldn't init worker");
 
-        assert_eq!(
-            true,
-            updater.update_ready().expect("couldn't check for update")
-        );
+        assert!(updater.update_ready().expect("couldn't check for update"));
         assert_eq!(VERSION_TEST, format!("{}", updater.current_version()));
     }
 
@@ -47,15 +44,14 @@ fn it_ignores_saved_version_after_an_upgrade_async() {
         assert_eq!(VERSION_TEST_NEW, format!("{}", updater.current_version()));
         updater.init().expect("couldn't init worker");
         // No more updates
-        assert_eq!(
-            false,
-            updater.update_ready().expect("couldn't check for update")
-        );
+        assert!(!updater.update_ready().expect("couldn't check for update"));
     }
 }
 
 #[test]
-#[should_panic(expected = "Error(Status(400), \"http://127.0.0.1:1234/releases/latest\")")]
+#[should_panic(
+    expected = "HTTP status client error (400 Bad Request) for url (http://127.0.0.1:1234/releases/latest)"
+)]
 fn it_handles_server_error_async() {
     setup_workflow_env_vars(true);
     first_check_after_installing_workflow();
@@ -79,32 +75,20 @@ fn it_caches_async_workers_payload() {
     // Next check will be immediate
     updater.set_interval(0);
     updater.init().expect("couldn't init worker");
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update"),
-    );
+    assert!(updater.update_ready().expect("couldn't check for update"),);
 
     // Consequent calls to update_ready should cache the payload.
     let _m = setup_mock_server(400);
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update"),
-    );
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update"),
-    );
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update"),
-    );
+    assert!(updater.update_ready().expect("couldn't check for update"),);
+    assert!(updater.update_ready().expect("couldn't check for update"),);
+    assert!(updater.update_ready().expect("couldn't check for update"),);
 
     {
         let mut updater = Updater::gh(MOCK_RELEASER_REPO_NAME).expect("cannot build Updater");
         // Next check will be immediate
         updater.set_interval(0);
         updater.init().expect("couldn't init worker");
-        assert_eq!(true, updater.update_ready().is_err());
+        assert!(updater.update_ready().is_err());
     }
 }
 
@@ -121,7 +105,9 @@ fn it_get_latest_info_from_releaser() {
         updater.set_interval(0);
         updater.init().expect("couldn't init worker");
 
-        assert!(updater.update_ready().expect("couldn't check for update"));
+        assert!(updater
+            .update_ready()
+            .expect("Blocking: couldn't check for update"));
     }
     {
         // Non-blocking
@@ -130,15 +116,16 @@ fn it_get_latest_info_from_releaser() {
         updater.set_interval(0);
         // Start async worker
         updater.init().expect("couldn't init worker");
-        let wait = time::Duration::from_millis(100);
+        let wait = time::Duration::from_millis(500);
         thread::sleep(wait);
 
         assert!(updater
             .try_update_ready()
-            .expect("couldn't check for update"));
+            .expect("Non-blocking: couldn't check for update"));
     }
 }
 
+#[allow(clippy::cast_possible_wrap)]
 #[test]
 fn it_does_one_network_call_per_interval() {
     {
@@ -166,7 +153,7 @@ fn it_does_one_network_call_per_interval() {
         let t = updater.update_ready();
         assert!(t.is_ok());
         // Make sure we still report update is ready
-        assert_eq!(true, t.unwrap());
+        assert!(t.unwrap());
 
         // Now we test that after interval has passed we will make a call
         let two_sec = time::Duration::from_secs(wait_time);
@@ -179,7 +166,7 @@ fn it_does_one_network_call_per_interval() {
 
             // Since server is returning error, update_ready() should fail.
             let t = updater.update_ready();
-            assert_eq!(true, t.is_err());
+            assert!(t.is_err());
         }
         {
             // Just making sure the next call will go through and return expected results.
@@ -192,11 +179,8 @@ fn it_does_one_network_call_per_interval() {
 
             // Since server is ok, update_ready() should work.
             let t = updater.update_ready();
-            assert_eq!(true, t.is_ok());
-            assert_eq!(
-                true,
-                updater.update_ready().expect("couldn't check for update")
-            );
+            assert!(t.is_ok());
+            assert!(updater.update_ready().expect("couldn't check for update"));
         }
     }
 }
@@ -216,10 +200,7 @@ fn it_tests_download() {
     updater.init().expect("couldn't init worker");
 
     // New update is available
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update")
-    );
+    assert!(updater.update_ready().expect("couldn't check for update"));
 
     let download_fn = updater.download_latest();
     assert!(download_fn.is_ok());
@@ -242,19 +223,16 @@ fn it_doesnt_download_without_release_info() {
     first_check_after_installing_workflow();
 
     let mut updater = Updater::gh(MOCK_RELEASER_REPO_NAME).expect("cannot build Updater");
-    updater.set_interval(864000);
+    updater.set_interval(864_000);
 
-    assert_eq!(false, updater.due_to_check());
+    assert!(!updater.due_to_check());
     updater.init().expect("couldn't init worker");
 
     assert!(updater.download_latest().is_err());
 
     // Since check time is due yet, following will just read cache without
     // getting any release info, hence the last line should panic
-    assert_eq!(
-        false,
-        updater.update_ready().expect("couldn't check for update")
-    );
+    assert!(!updater.update_ready().expect("couldn't check for update"));
     updater.download_latest().unwrap();
 }
 
@@ -269,10 +247,7 @@ fn it_downloads_after_getting_release_info() {
     updater.init().expect("couldn't init worker");
     assert!(updater.download_latest().is_err());
 
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update")
-    );
+    assert!(updater.update_ready().expect("couldn't check for update"));
     assert!(updater.download_latest().is_ok());
 }
 
@@ -287,10 +262,7 @@ fn it_tests_async_updates_1() {
     // Next check will be immediate
     updater.set_interval(0);
     updater.init().expect("couldn't init worker");
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update")
-    );
+    assert!(updater.update_ready().expect("couldn't check for update"));
 }
 
 #[test]
@@ -314,10 +286,7 @@ fn it_tests_async_updates_2() {
     // Increase interval
     updater.set_interval(86400);
 
-    assert_eq!(
-        true,
-        updater.update_ready().expect("couldn't check for update")
-    );
+    assert!(updater.update_ready().expect("couldn't check for update"));
 }
 
 pub(super) fn setup_workflow_env_vars(secure_temp_dir: bool) -> PathBuf {
@@ -358,8 +327,5 @@ fn first_check_after_installing_workflow() {
     updater.init().expect("couldn't init worker");
 
     // First update_ready is always false.
-    assert_eq!(
-        false,
-        updater.update_ready().expect("couldn't check for update")
-    );
+    assert!(!updater.update_ready().expect("couldn't check for update"));
 }